+++ intmath/intmath.go
