+++ utils/math/BUILD.bazel
@@ -16,7 +16,7 @@ go_library(
     deps = [
         "//ids",
         "//utils/heap",
-        "//vms/saevm/intmath",
+        "//intmath",
         "@org_golang_x_exp//constraints",
     ],
 )
