+++ vms/platformvm/txs/executor/proposal_tx_executor_test.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/ava-labs/avalanchego/database"
 	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/avalanchego/intmath"
 	"github.com/ava-labs/avalanchego/upgrade/upgradetest"
 	"github.com/ava-labs/avalanchego/utils/crypto/secp256k1"
 	"github.com/ava-labs/avalanchego/utils/hashing"
@@ -958,5 +959,109 @@ func TestProposalTxExecuteAddValidator(t *testing.T) {
 			onAbortState,
 		)
 		require.ErrorIs(err, ErrFlowCheckFailed)
+	}
+}
+
+func TestMulDiv(t *testing.T) {
+	tests := []struct {
+		name    string
+		a       uint64
+		b       uint64
+		c       uint64
+		want    uint64
+		wantErr error
+	}{
+		{
+			name:    "division_by_zero",
+			a:       100,
+			b:       5,
+			c:       0,
+			wantErr: errDivideByZero,
+		},
+		{
+			name: "a_is_zero",
+			a:    0,
+			b:    4,
+			c:    50,
+			want: 0,
+		},
+		{
+			name: "b_is_zero",
+			a:    250,
+			b:    0,
+			c:    50,
+			want: 0,
+		},
+		{
+			name: "basic_case_1",
+			a:    100,
+			b:    3,
+			c:    10,
+			want: 30,
+		},
+		{
+			name: "basic_case_2",
+			a:    250,
+			b:    4,
+			c:    50,
+			want: 20,
+		},
+		{
+			name: "precision",
+			a:    7,
+			b:    3,
+			c:    10,
+			want: 2,
+		},
+		{
+			name:    "overflow",
+			a:       math.MaxUint64,
+			b:       10,
+			c:       2,
+			wantErr: intmath.ErrOverflow,
+		},
+		{
+			name: "round_down",
+			a:    10,
+			b:    10,
+			c:    30,
+			want: 3,
+		},
+		{
+			name: "round_up",
+			a:    20,
+			b:    10,
+			c:    30,
+			want: 7,
+		},
+		{
+			name: "large_values_without_overflow",
+			a:    300_000_000_000,
+			b:    200_000_000_000,
+			c:    400_000_000_000,
+			want: 150_000_000_000,
+		},
+		{
+			name: "small_a_large_c",
+			a:    5,
+			b:    3,
+			c:    10,
+			want: 2,
+		},
+		{
+			name: "maxUint64_*_maxUint64_/_maxUint64",
+			a:    math.MaxUint64,
+			b:    math.MaxUint64,
+			c:    math.MaxUint64,
+			want: math.MaxUint64,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := mulDivRound(tt.a, tt.b, tt.c)
+			require.ErrorIs(t, err, tt.wantErr)
+			require.Equal(t, tt.want, got)
+		})
 	}
 }
