+++ vms/saevm/gasprice/estimator.go
@@ -22,9 +22,9 @@ import (
 	"github.com/ava-labs/libevm/rpc"
 	"go.uber.org/zap"
 
+	"github.com/ava-labs/avalanchego/intmath"
 	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/ava-labs/avalanchego/vms/saevm/blocks"
-	"github.com/ava-labs/avalanchego/vms/saevm/intmath"
 )
 
 // Backend that the [Estimator] depends on for chain data.
