+++ vms/saevm/gastime/acp176.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/holiman/uint256"
 
+	"github.com/ava-labs/avalanchego/intmath"
 	"github.com/ava-labs/avalanchego/vms/components/gas"
-	"github.com/ava-labs/avalanchego/vms/saevm/intmath"
 )
 
 // BeforeBlock is intended to be called before processing a block with the
