+++ vms/saevm/gastime/gastime.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/holiman/uint256"
 
+	"github.com/ava-labs/avalanchego/intmath"
 	"github.com/ava-labs/avalanchego/vms/components/gas"
-	"github.com/ava-labs/avalanchego/vms/saevm/intmath"
 	"github.com/ava-labs/avalanchego/vms/saevm/proxytime"
 )
