+++ vms/saevm/proxytime/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
     importpath = "github.com/ava-labs/avalanchego/vms/saevm/proxytime",
     visibility = ["//visibility:public"],
     deps = [
-        "//vms/saevm/intmath",
+        "//intmath",
         "@com_github_google_go_cmp//cmp",
         "@com_github_google_go_cmp//cmp/cmpopts",
         "@com_github_stephenbuttolph_canoto//:canoto",
