+++ vms/saevm/proxytime/proxytime.go
@@ -12,7 +12,7 @@ import (
 	"math/bits"
 	"time"
 
-	"github.com/ava-labs/avalanchego/vms/saevm/intmath"
+	"github.com/ava-labs/avalanchego/intmath"
 )
 
 //go:generate go run github.com/StephenButtolph/canoto/canoto $GOFILE
