+++ vms/saevm/sae/BUILD.bazel
@@ -190,7 +190,7 @@ go_test(
         "//vms/saevm/cmputils",
         "//vms/saevm/gastime",
         "//vms/saevm/hook/hookstest",
-        "//vms/saevm/intmath",
+        "//intmath",
         "//vms/saevm/params",
         "//vms/saevm/sae/rpc",
         "//vms/saevm/saedb",
